Changed Point to use doubles instead of ints, and added functions to make it usable as a vector (e.g. get_magnitude(), get_angle(), etc.). Added getters/setters to Player that use Points. Added a couple not-yet-used functions to Player for adjusting velocity (bounce() and stop()).
git-svn-id: https://legesmotus.svn.sourceforge.net/svnroot/legesmotus/legesmotus/trunk@631 a5a5d5d7-22d1-44d1-b37a-67d3d99fcfd7

+++ client/GraphicalMap.cpp
@@ -100,8 +100,8 @@ void	GraphicalMap::add_object(MapReader& object_data) {
 
 			bool		is_tiled = flags.find('T') != string::npos;
 			bool		is_obstacle = flags.find('O') != string::npos;
-			int		width = 0;
-			int		height = 0;
+			double		width = 0;
+			double		height = 0;
 
 			if (is_tiled) {
 				// Tiled graphic
@@ -122,8 +122,8 @@ void	GraphicalMap::add_object(MapReader& object_data) {
 				}
 			} else {
 				Sprite*	sprite = load_graphic<Sprite>(sprite_name);
-				width = int(sprite->get_image_width());
-				height = int(sprite->get_image_height());
+				width = sprite->get_image_width();
+				height = sprite->get_image_height();
 				map_object.set_sprite(sprite);
 			}
 
@@ -184,7 +184,7 @@ void	GraphicalMap::add_object(MapReader& object_data) {
 			sprite->set_antialiasing(false);
 			map_object.set_sprite(sprite);
 
-			map_object.get_bounding_polygon().make_rectangle(int(sprite->get_image_width()) + GATE_EXTENT * 2, GATE_HEIGHT, upper_left - Point(GATE_EXTENT, 0)); // XXX: casting double to int here
+			map_object.get_bounding_polygon().make_rectangle(sprite->get_image_width() + GATE_EXTENT * 2, GATE_HEIGHT, upper_left - Point(GATE_EXTENT, 0));
 		}
 
 		break;
